fix: reduce styles_writer.rs and text_writer.rs to fit 300-line budget
- styles_writer.rs (302 → 288): replace three thin private wrapper
  functions with direct imports from styles_utils
- text_writer.rs (340 → 259): move multi-page and overflow pagination
  tests into a new focused module text_writer_pages.rs (142 lines)
